Initial refactoring of nearby fragment
